requirements.txt
pandas 
notebook
numpy
scikit-learn
xgboost
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-pyYAML
Flask
Flask-Cors
-e .